% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitToken_check.R
\name{bitToken_check}
\alias{bitToken_check}
\title{Check if a dataset is relevant for tokenizing with bitToken()}
\usage{
bitToken_check(data, text_column)
}
\arguments{
\item{data}{A data frame to check for tokenization}

\item{text_column}{The name of the column in \code{data} to check for tokenization}
}
\value{
A message indicating whether the specified text column contains only text data, or an error message if the column contains non-text data.
}
\description{
This function checks if a dataset is relevant for tokenizing with bitToken() by ensuring that the specified text column contains only text data.
}
\examples{
\dontrun{
library(bitToken2)
data(mtcars)
bitToken_check(mtcars, "hp")
data(chatGPT_news1)
bitToken_check(chatGPT_news1, "title")
}

}
