% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitToken_count.R
\name{bitToken_count}
\alias{bitToken_count}
\title{Token Count}
\usage{
bitToken_count(data, text_column, pattern, location = FALSE, sum_info = FALSE)
}
\arguments{
\item{data}{A data frame containing the text data.}

\item{text_column}{A string specifying the name of the text column in the input data.}

\item{pattern}{A character string representing the specific character or string to count.
Note the following when using the pattern argument:
- Case sensitivity: By default, the function is case sensitive. Use (?i) within the pattern to make it case insensitive.
- Regular expressions: The function uses regular expressions, so escape special characters (e.g., . or *) with a backslash (\\).
- Spaces and whitespace: Include the exact whitespace character in the pattern, such as regular spaces or tabs, or use \\s to represent any whitespace.}

\item{location}{A boolean flag indicating whether to return the positions of tokens containing the specific pattern. Default is FALSE.}

\item{sum_info}{A boolean flag indicating whether to return summary information and frequency table for counts. Default is FALSE.}
}
\value{
A list or a numeric vector with the count of occurrences of the specific character or string for each row, and additional information based on the user options.
}
\description{
Count the occurrences of a specific character or string in each row of the specified text column in a data frame.
}
\examples{
chatGPT_news1$comma_count <- bitToken_count(chatGPT_news1, "title", pattern = ",")

}
