% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitToken_info.R
\name{bitToken_info}
\alias{bitToken_info}
\title{Token Information}
\usage{
bitToken_info(data, text_column, add = FALSE, min_t = 1, max_t = Inf)
}
\arguments{
\item{data}{A character vector or a list of character vectors.}

\item{text_column}{A string specifying the name of the text column in the input data.}

\item{add}{A logical indicating whether to include additional information about the data.}

\item{min_t}{The minimum number of tokens to consider for summary statistics.}

\item{max_t}{The maximum number of tokens to consider for summary statistics.}
}
\value{
A named numeric vector with the minimum, median, mode, and maximum number of tokens.
}
\description{
Get information about the number of tokens in a character vector or list of character vectors.
}
\examples{
data("chatGPT_news1", package = "bitToken2")
bitToken_info(chatGPT_news1, "title")
bitToken_info(chatGPT_news1, "title", min_t = 3, max_t = 10, add = TRUE)

}
