% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitToken_location.R
\name{bitToken_location}
\alias{bitToken_location}
\title{Locate specific pattern in a text column of a data frame}
\usage{
bitToken_location(data, text_column, pattern)
}
\arguments{
\item{data}{Data frame that contains the text column.}

\item{text_column}{Name of the column in the data frame that contains the text to search for the pattern.}

\item{pattern}{Pattern to locate in the text.}
}
\value{
A list of vectors indicating the presence of the pattern in each token. The length of the vector is equal to the number of tokens in the corresponding row.
        Each element in the vector is either 0 (the pattern is not present) or 1 (the pattern is present).
}
\description{
This function locates a specific pattern in a text column of a data frame.
It returns a binary vector indicating the presence of the pattern in each token.
}
\examples{
data(chatGPT_news1, package="bitToken2")
bitToken_location(chatGPT_news1, "title", "COVID")
}
