% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitToken_min_m.R
\name{bitToken_min_m}
\alias{bitToken_min_m}
\title{Calculate the number of tokens in a given text column and find rows with at least a specified number of tokens and not more than a maximum number (Multicore)}
\usage{
bitToken_min_m(
  data,
  col_name,
  min_tokens = 3,
  max_tokens = NULL,
  value = FALSE,
  use_p = TRUE,
  num_cores = parallel::detectCores()
)
}
\arguments{
\item{data}{a dataframe}

\item{col_name}{the name of the text column}

\item{min_tokens}{the minimum number of tokens expected in the text column (default = 3)}

\item{max_tokens}{the maximum number of tokens expected in the text column (default = NULL)}

\item{value}{a boolean value. If TRUE, the function returns the actual row values instead of the row indices (default = FALSE)}

\item{use_p}{A logical value. If \code{TRUE}, the function will use parallel processing. Default is \code{TRUE}.}

\item{num_cores}{The number of cores to be used for parallel processing. The default is the half of the total number of cores available in the system.}
}
\value{
A vector of either the row indices or the actual row values based on the 'value' option
}
\description{
This function takes in a dataframe, the name of the text column and the minimum number of tokens expected in the text column. It returns either the row indices or the actual row values based on the 'value' option
}
\examples{
tokens5_to_7 <- bitToken_min_m(
                chatGPT_news1, "title", min_tokens = 5,
                max_tokens = 7, value = TRUE, use_p = FALSE
                )

}
