% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitToken_order.R
\name{bitToken_order}
\alias{bitToken_order}
\title{Get Tokens in Descending Order by Frequency}
\usage{
bitToken_order(data, text_column, token_num)
}
\arguments{
\item{data}{A data frame containing the text column.}

\item{text_column}{The name of the column containing the text strings.}

\item{token_num}{The index of the token to extract (starting from 1).}
}
\value{
A data frame with two columns: "Token" and "Frequency", sorted by "Frequency" in descending order.
}
\description{
Extracts the specified token number from a column of text strings and returns a vector
of the tokens in descending order by their frequency.
}
\examples{
bitToken_order(chatGPT_news1, "title", 1)

}
