% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitToken_order_m.R
\name{bitToken_order_m}
\alias{bitToken_order_m}
\title{Get Tokens in Descending Order by Frequency (Multicore)}
\usage{
bitToken_order_m(
  data,
  text_column,
  token_num,
  use_p = TRUE,
  num_cores = parallel::detectCores()/2
)
}
\arguments{
\item{data}{A data frame containing the text column.}

\item{text_column}{The name of the column containing the text strings.}

\item{token_num}{The index of the token to extract (starting from 1).}

\item{use_p}{A logical value. If \code{FALSE}, the function will not use the 'use_p' argument in its calculations. Default is \code{TRUE}.}

\item{num_cores}{The number of cores to be used for parallel processing. The default is the half of the total number of cores available in the system.}
}
\value{
A data frame with two columns: "Token" and "Frequency", sorted by "Frequency" in descending order.
}
\description{
Extracts the specified token number from a column of text strings and returns a vector
of the tokens in descending order by their frequency.
}
\examples{
bitToken_order_m(chatGPT_news1, "title", 1, use_p = FALSE)

}
