% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitToken_pattern_trend.R
\name{bitToken_pattern_trend}
\alias{bitToken_pattern_trend}
\title{Compute text pattern trend over time}
\usage{
bitToken_pattern_trend(
  data,
  text_column,
  pattern,
  date_column = NULL,
  window = NULL
)
}
\arguments{
\item{data}{a data frame or a data.table}

\item{text_column}{a character string specifying the name of the text column in the data frame}

\item{pattern}{a character string specifying the pattern to search for in the text column}

\item{date_column}{a character string specifying the name of the date column in the data frame (optional)}

\item{window}{an integer specifying the moving average window (optional)}
}
\value{
A data.table with columns date, total, count, percentage, occurrence, occurrence percentage, and occurrence per count, as well as their moving averages if a window is given. If a date column is not given, a data.table with one row and the same columns (except date) is returned.
}
\description{
This function takes a dataframe, a text column name, a pattern to search for, and optionally a date column name and a moving average window. It computes the count, percentage, occurrence, occurrence percentage, and occurrence per count of the pattern in the text column on a daily basis if a date column is given, and then computes their moving averages if a window is given. It finally plots these trend lines over time.
}
\examples{
\dontrun{
bitToken_pattern_trend(data = data, text_column = "text", pattern = "bitcoin", date_column = "date", window = 7)
}
}
