% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitToken_pattern_trend_viz.R
\name{bitToken_pattern_trend_viz}
\alias{bitToken_pattern_trend_viz}
\title{Plot Count and Moving Average Count Over Time}
\usage{
bitToken_pattern_trend_viz(df, x_var, y_var1, y_var2, window = NULL)
}
\arguments{
\item{df}{A data frame.}

\item{x_var}{A character string specifying the name of the x variable (should be a Date column).}

\item{y_var1}{A character string specifying the name of the first y variable (Count).}

\item{y_var2}{A character string specifying the name of the second y variable (Moving Average Count).}

\item{window}{A numeric value specifying the window size for calculating moving averages. If NULL or greater than the number of observations in df, no moving averages are calculated.}
}
\value{
A ggplot object.
}
\description{
This function takes a data frame and the names of two count variables, plots these variables over time, and returns the resulting ggplot object.
}
\examples{
\dontrun{
bitToken_pattern_trend_viz(df = df, x_var = "date", y_var1 = "count", y_var2 = "count_ma", window = 7)
}
}
