% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitToken_search.R
\name{bitToken_search}
\alias{bitToken_search}
\title{Token Search}
\usage{
bitToken_search(data, text_column, pattern, location = NULL)
}
\arguments{
\item{data}{A data frame containing the text data.}

\item{text_column}{A string specifying the name of the text column in the input data.}

\item{pattern}{A character string representing the specific character or string to filter.
Note the following when using the pattern argument:
- Case sensitivity: By default, the function is case sensitive. Use (?i) within the pattern to make it case insensitive.
- Regular expressions: The function uses regular expressions, so escape special characters (e.g., . or *) with a backslash (\\).
- Spaces and whitespace: Include the exact whitespace character in the pattern, such as regular spaces or tabs, or use \\s to represent any whitespace.}

\item{location}{A numeric value indicating the position of tokens to consider when filtering. Default is NULL, indicating all tokens are considered.}
}
\value{
A character vector with the filtered values from the specified text column.
}
\description{
Filter rows based on the occurrence of a specific character or string at a specific location in the specified text column in a data frame.
}
\examples{
# Search for 'COVID' at second position in titles
titles_with_COVID <- bitToken_search(chatGPT_news1, "title", "COVID", location=2)

}
