% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitToken_search2.R
\name{bitToken_search2}
\alias{bitToken_search2}
\title{Bit Token Search 2}
\usage{
bitToken_search2(data, text_column, pattern, index = FALSE)
}
\arguments{
\item{data}{A data frame containing the text data.}

\item{text_column}{A string specifying the name of the text column in the input data.}

\item{pattern}{A character string representing the specific character or string to filter.}

\item{index}{A logical value indicating whether to return the index of the original rows. 
Default is FALSE. If TRUE, the function returns a list of lists, 
each containing the index of the original row and the filtered sentences from that row.}
}
\value{
If index = FALSE, a character vector with the filtered sentences.
  If index = TRUE, a list of lists, each containing the index of the original row and the filtered sentences from that row.
}
\description{
This function searches for a specific word in a paragraph style text and returns the sentences containing the word. 
The function can optionally return the index of the original rows.
}
\examples{
# Search for 'COVID' in text column without index
sentences_COVID <- bitToken_search2(chatGPT_news1, "title", "COVID")
# Search for 'COVID' in text column with index
sentences_COVID_index <- bitToken_search2(chatGPT_news1, "title", "COVID", index = TRUE)
}
