% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitToken_viz.R
\name{bitToken_viz}
\alias{bitToken_viz}
\title{Visualize Token Information for Character Vectors}
\usage{
bitToken_viz(
  x,
  type = "boxplot",
  round = FALSE,
  remove_outliers = FALSE,
  info = FALSE
)
}
\arguments{
\item{x}{a character vector or a list of character vectors.}

\item{type}{the type of plot to generate. Possible values are "boxplot" (default) and "histogram".}

\item{round}{an optional logical value indicating whether to round the output to integers.
The default value is FALSE.}

\item{remove_outliers}{an optional logical value indicating whether to remove outliers from the boxplot.
The default value is FALSE.}

\item{info}{an optional logical value indicating whether to print the median and mean values on the plot.
The default value is FALSE.}
}
\value{
A plot of the token information for the input character vectors.
}
\description{
This function takes in a character vector or a list of character vectors and generates
a plot of their token information. The default plot type is a boxplot, but a histogram
plot is also available.
}
\examples{
bitToken_viz(chatGPT_news1$title, type = "boxplot", remove_outliers = TRUE, info = TRUE)
bitToken_viz(chatGPT_news1$title, type = "histogram", info = TRUE)

}
