% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitToken_viz_m.R
\name{bitToken_viz_m}
\alias{bitToken_viz_m}
\title{Visualize Token Information for Character Vectors}
\usage{
bitToken_viz_m(
  data,
  text_column,
  type = "boxplot",
  round = FALSE,
  rm_outliers = FALSE,
  info = FALSE,
  use_p = TRUE,
  plot_title = "Token Information Visualization"
)
}
\arguments{
\item{data}{A data frame.}

\item{text_column}{The name of the column that contains the text data.}

\item{type}{The type of plot to generate. Possible values are "boxplot" (default) and "histogram".}

\item{round}{An optional logical value indicating whether to round the output to integers.
The default value is FALSE.}

\item{rm_outliers}{An optional logical value indicating whether to remove outliers from the boxplot.
The default value is FALSE.}

\item{info}{An optional logical value indicating whether to print the median and mean values on the plot.
The default value is FALSE.}

\item{use_p}{A logical value. If \code{FALSE}, the function will not use the 'use_p' argument in its calculations. Default is \code{TRUE}.}

\item{plot_title}{An optional character string for the plot title.
The default value is "Token Information Visualization".}
}
\value{
A plot of the token information for the input character vectors.
}
\description{
This function takes in a character vector or a list of character vectors and generates
a plot of their token information. The default plot type is a boxplot, but a histogram
plot is also available.
}
\examples{
bitToken_viz_m(chatGPT_news1, "title", rm_outliers = TRUE, info = TRUE, use_p = FALSE)
bitToken_viz_m(chatGPT_news1, "title", type = "histogram", info = TRUE, use_p = FALSE)

}
