% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatGPT_news1.R
\docType{data}
\name{chatGPT_news1}
\alias{chatGPT_news1}
\title{Sample news article titles and categories}
\format{
A data frame with 30 rows and 2 variables:
\describe{
  \item{title}{A character vector of news article titles}
  \item{category}{A character vector of news article categories}
}
}
\usage{
chatGPT_news1
}
\description{
A dataset containing 30 news article titles and their corresponding categories
}
\examples{
# Load the dataset
data("chatGPT_news1")

# View the first 6 rows of the dataset
head(chatGPT_news1)

}
\keyword{datasets}
