% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitToken_m.R
\name{bitToken_m}
\alias{bitToken_m}
\title{Tokenize text data in a data frame (Multicore)}
\usage{
bitToken_m(
  data,
  text_column,
  filter_var = NULL,
  filter_vals = NULL,
  lengths = FALSE,
  use_p = TRUE,
  num_cores = parallel::detectCores()
)
}
\arguments{
\item{data}{A data frame containing the text data to tokenize}

\item{text_column}{The name of the column in \code{data} containing the text data to tokenize}

\item{filter_var}{The name of a column in \code{data} to filter on (optional)}

\item{filter_vals}{A vector of values to filter on in \code{filter_var} (optional)}

\item{lengths}{A logical value indicating whether to return the lengths of the tokens (default = FALSE)}

\item{use_p}{A logical value. If \code{FALSE}, the function will not use the 'use_p' argument in its calculations. Default is \code{TRUE}.}

\item{num_cores}{The number of cores to use for parallel processing.}
}
\value{
A list of token vectors or a vector of token lengths
}
\description{
Tokenize text data in a data frame (Multicore)
}
