% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitToken_min.R
\name{bitToken_min}
\alias{bitToken_min}
\title{Calculate the number of tokens in a given text column and find rows with at least a specified number of tokens}
\usage{
bitToken_min(data, col_name, min_tokens = 3, value = FALSE)
}
\arguments{
\item{data}{a dataframe}

\item{col_name}{the name of the text column}

\item{min_tokens}{the minimum number of tokens expected in the text column (default = 3)}

\item{value}{a boolean value. If TRUE, the function returns the actual row values instead of the row indices (default = FALSE)}
}
\value{
A vector of either the row indices or the actual row values based on the 'value' option
}
\description{
This function takes in a dataframe, the name of the text column and the minimum number of tokens expected in the text column. It returns either the row indices or the actual row values based on the 'value' option
}
\examples{
bitToken_min(chatGPT_news1, "title", min_tokens = 5, value = TRUE)

}
