% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatGPT_news1.R
\docType{data}
\name{chatGPT_news1}
\alias{chatGPT_news1}
\title{Sample news article titles and categories}
\format{
An object of class \code{data.frame} with 30 rows and 2 columns.
}
\usage{
chatGPT_news1
}
\description{
A dataset containing 30 news article titles and their corresponding categories
}
\keyword{datasets}
