% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chatGPT_news1}
\alias{chatGPT_news1}
\alias{chatGPT_df1}
\title{News dataset for bitToken2 package}
\format{
An object of class \code{data.frame} with 20 rows and 2 columns.
}
\usage{
chatGPT_df1
}
\description{
A dataset containing 20 rows of news article titles and categories, designed for use with the bitToken2 package.
}
\keyword{datasets}
